refactor cruaks to separate aksConfig
